+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.0-preview.5 (WIP)
 **Bug Fixes**
-- 
+- [#968](https://github.com/Azure/azure-storage-fuse/issues/968) Duplicate directory listing
+- [#964](https://github.com/Azure/azure-storage-fuse/issues/964) Rename for FNS account failing with source does not exists error
 
 ## 2.0.0-preview.4 (2022-11-03)
 **Breaking Changes**