+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.0-preview.5 (WIP)
 **Bug Fixes**
-- [#968](https://github.com/Azure/azure-storage-fuse/issues/968) Sometimes a directory would list twice for FNS account
+- Fixed a bug where sometimes a directory would list twice
 - [#972](https://github.com/Azure/azure-storage-fuse/issues/972) Mount all fails 
 
 ## 2.0.0-preview.4 (2022-11-03)