+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 **Bug Fixes**
 - [#968](https://github.com/Azure/azure-storage-fuse/issues/968) Duplicate directory listing
 - [#964](https://github.com/Azure/azure-storage-fuse/issues/964) Rename for FNS account failing with source does not exists error
+- [#972](https://github.com/Azure/azure-storage-fuse/issues/972) Mount all fails 
 - Added support for "-o nonempty" to mount on a non-empty mount path
 
 **Breaking Changes**